+++ Sources/TuistKit/Commands/DumpCommand.swift
@@ -8,7 +8,7 @@ struct DumpCommand: ParsableCommand {
     static var configuration: CommandConfiguration {
         CommandConfiguration(
             commandName: "dump",
-            abstract: "Outputs the project manifest as a JSON"
+            abstract: "Outputs the manifest as a JSON"
         )
     }
 
@@ -16,12 +16,17 @@ struct DumpCommand: ParsableCommand {
 
     @Option(
         name: .shortAndLong,
-        help: "The path to the folder where the project manifest is",
+        help: "The path to the folder where the manifest is",
         completion: .directory
     )
     var path: String?
 
+    @Argument(help: "The manifest to be dumped")
+    var manifest: DumpableManifest = .project
+
     func run() throws {
-        try DumpService().run(path: path)
+        try DumpService().run(path: path, manifest: manifest)
     }
 }
+
+extension DumpableManifest: ExpressibleByArgument {}